Move FrameData out of VulkanRenderManager
