+++ Common/GPU/Vulkan/VulkanRenderManager.h
@@ -65,11 +65,6 @@ private:
 	std::string tag_;
 };
 
-enum class VKRRunType {
-	END,
-	SYNC,
-};
-
 enum {
 	MAX_TIMESTAMP_QUERIES = 128,
 };
@@ -486,45 +481,6 @@ private:
 	// Permanent objects
 	VkSemaphore acquireSemaphore_;
 	VkSemaphore renderingCompleteSemaphore_;
-
-	// Per-frame data, round-robin so we can overlap submission with execution of the previous frame.
-	struct FrameData {
-		std::mutex push_mutex;
-		std::condition_variable push_condVar;
-
-		std::mutex pull_mutex;
-		std::condition_variable pull_condVar;
-
-		bool readyForFence = true;
-		bool readyForRun = false;
-		bool skipSwap = false;
-		VKRRunType type = VKRRunType::END;
-
-		VkFence fence;
-		VkFence readbackFence;  // Strictly speaking we might only need one of these.
-		bool readbackFenceUsed = false;
-
-		// These are on different threads so need separate pools.
-		VkCommandPool cmdPoolInit;  // Written to from main thread
-		VkCommandPool cmdPoolMain;  // Written to from render thread, which also submits
-
-		VkCommandBuffer initCmd;
-		VkCommandBuffer mainCmd;
-		VkCommandBuffer presentCmd;
-
-		bool hasInitCommands = false;
-		bool hasPresentCommands = false;
-
-		std::vector<VKRStep *> steps;
-
-		// Swapchain.
-		bool hasBegun = false;
-		uint32_t curSwapchainImage = -1;
-
-		// Profiling.
-		QueueProfileContext profile;
-		bool profilingEnabled_;
-	};
 
 	FrameData frameData_[VulkanContext::MAX_INFLIGHT_FRAMES];
 	int newInflightFrames_ = -1;
